validate nut xoa post

+++ resources/views/Goodi/Submission/show.blade.php
@@ -297,7 +297,9 @@
                                         <ul>
                                             <li><a href="/idea/show/{{ $idea->id }}">Open Idea</a></li>
                                             <li><a href="">Change Content</a></li>
-                                            <li><a href="/idea/delete/{{ $idea->id }}">Remove Post</a></li>
+                                            @if ($idea->id == Auth::user()->id || Auth::user()->id == 1)
+                                                <li><a href="/idea/delete/{{ $idea->id }}">Remove Post</a></li>
+                                            @endif
                                         </ul>
                                     </div>
                                 </div>
